requirements.txt
nltk>=3.8.1
scikit-learn>=1.4.0
rank-bm25>=0.2.2
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0.1
